requirements.txt
click==8.1.7
DrissionPage==4.0.5.6
DrissionPage==4.1.0.12
pycryptodome==3.20.0
Requests==2.32.3
